+++ src/pages/RestaurantListPage/RestaurantList.ts
@@ -1,7 +1,7 @@
-import type { Component } from '../interface';
-import type { Category, Restaurant, SortBy } from '../type';
-import RestaurantListItem from './RestaurantListItem';
-import { DEFAULT_CATEGORY } from '../utils/constants';
+import type { Component } from '../../interface';
+import type { Category, Restaurant, SortBy } from '../../type';
+import RestaurantListItem from '../../components/RestaurantListItem';
+import { DEFAULT_CATEGORY } from '../../utils/constants';
 
 type RestaurantListState = {
   restaurants: Restaurant[];
