+++ src/pages/RestaurantListPage/index.ts
@@ -1,6 +1,6 @@
 import type { Component } from '../../interface';
 import type { Category, SortBy, Restaurant } from '../../type';
-import RestaurantList from '../../components/RestaurantList';
+import RestaurantList from './RestaurantList';
 import RestaurantFilterContainer from './RestaurantFilterContainer';
 import { DEFAULT_CATEGORY, REQUEST_RASTAURANT_KEY } from '../../utils/constants';
 import GNB from '../../components/GNB';
